WIP: split stripe/square test files — NOT GREEN, jscpd finds 71 clones to fix
