+++ biome.json
@@ -156,9 +156,7 @@
         "src/shared/db/migrations/schema.ts",
         "src/ui/templates/public/reservations.tsx",
         "test/lib/server-listing-parents.test.ts",
-        "test/lib/server-payments.test.ts",
-        "test/lib/square.test.ts",
-        "test/lib/stripe.test.ts"
+        "test/lib/server-payments.test.ts"
       ],
       "linter": {
         "rules": {
